Move mosaicStyles into GlobalCss
